uniclub-api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg[binary]
python-dotenv
pydantic
pydantic-settings
email-validator
alembic
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
